+++ src/components/app/AppToolbar.tsx
@@ -39,6 +39,7 @@ import useWindowFocus from "ui/hooks/use-window-focus";
 import useWindowSize from "ui/hooks/use-window-size";
 import { useAppDispatch, useAppSelector } from "store/hooks";
 import API from "renderer/lib/api";
+import clipboardActions from "store/features/clipboard/clipboardActions";
 
 const sectionAccelerators = {
   world: "CommandOrControl+1",
@@ -94,6 +95,10 @@ const AppToolbar: FC = () => {
     }),
     [],
   );
+
+  useEffect(() => {
+    dispatch(clipboardActions.fetchClipboard());
+  }, [dispatch]);
 
   const onRun = useCallback(() => {
     dispatch(buildGameActions.buildGame({ buildType: "web" }));
@@ -175,7 +180,7 @@ const AppToolbar: FC = () => {
     if (e.target && (e.target as Node).nodeName !== "BODY") {
       return;
     }
-    if (e.key === "/" && searchInputRef.current) {
+    if (!e.metaKey && !e.ctrlKey && e.key === "/" && searchInputRef.current) {
       searchInputRef.current.focus();
       e.preventDefault();
     }