+++ src/components/ui/menu/ContextMenu.tsx
@@ -1,4 +1,4 @@
-import React, { useEffect, useRef } from "react";
+import React, { useCallback, useEffect, useRef } from "react";
 import useNestedMenu from "ui/hooks/use-nested-menu";
 import { PositionedPortal } from "ui/layout/PositionedPortal";
 import { Menu } from "ui/menu/Menu";
@@ -8,13 +8,21 @@ interface ContextMenuProps {
   readonly y: number;
   readonly children?: JSX.Element[];
   readonly onClose?: () => void;
+  readonly onKeyDown?: (e: React.KeyboardEvent) => boolean;
 }
 
-export const ContextMenu = ({ x, y, children, onClose }: ContextMenuProps) => {
+export const ContextMenu = ({
+  x,
+  y,
+  children,
+  onClose,
+  onKeyDown,
+}: ContextMenuProps) => {
   const { menuRef, isOpen, childrenWithProps } = useNestedMenu(
     children,
     true,
     "right",
+    onKeyDown,
   );
   const wasOpen = useRef(isOpen);
   useEffect(() => {
@@ -23,6 +31,20 @@ export const ContextMenu = ({ x, y, children, onClose }: ContextMenuProps) => {
     }
     wasOpen.current = isOpen;
   }, [isOpen, onClose]);
+
+  const onHotkeyHandled = useCallback(() => {
+    onClose?.();
+  }, [onClose]);
+
+  useEffect(() => {
+    if (isOpen) {
+      document.addEventListener("gbs::hotkey", onHotkeyHandled);
+    }
+    return () => {
+      document.removeEventListener("gbs::hotkey", onHotkeyHandled);
+    };
+  }, [isOpen, onHotkeyHandled]);
+
   return isOpen ? (
     <PositionedPortal x={x} y={y} offsetX={-2} offsetY={-10}>
       <Menu ref={menuRef}>{childrenWithProps}</Menu>