+++ src/components/ui/menu/style.ts
@@ -5,7 +5,7 @@ interface StyledMenuAcceleratorProps {
   $accelerator: string;
 }
 
-const acceleratorForPlatform = (accelerator: string) => {
+export const acceleratorForPlatform = (accelerator: string) => {
   if (API.platform === "darwin") {
     return accelerator
       .replace(/CommandOrControl\+/g, "⌘")